requirements.txt
streamlit
langchain
openai
opencensus-ext-azure